pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fastq_read_stats"
dynamic = ["version"]
authors = [{ name = "Ashley Shalloe", email = "ashley.shalloe@ukhsa.gov.uk" }]
description = "Generates basic statistics given a fastq.gz as input"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pyfunctional>=1.5.0", "pytest>=8.2.2", "pyfastx>=2.2.0"]

[project.urls]
Homepage = "https://gitlab.phe.gov.uk/gpha/mscape/fastq_read_stats"
Issues = "https://gitlab.phe.gov.uk/gpha/mscape/fastq_read_stats/-/issues"

[project.scripts]
fastq_read_stats = "fastq_read_stats.fastq_read_stats:main"
fastq_read_statsMP = "fastq_read_stats.fastq_read_statsMP:main"

[tool.setuptools.dynamic]
version = { attr = "fastq_read_stats.fastq_read_stats.__version__" }
